+++ src/exports/client.ts
@@ -1,3 +1,4 @@
 export { LockTranslation } from '../components/LockTranslation/index.js'
+export { OpenAiModelField } from '../components/OpenAiModelField.js'
 export { TranslationControl } from '../components/TranslationControl.js'
 export type * from '../types/index.js'